Fix service dependencies
These were copied from a [DO tutorial][1], and are incorrect.  With
`Requires=etcd2`, if etcd2 hiccups and restarts, then all the
containers on the system restart, too.  That's a very bad thing.

Instead, use `Wants=etcd2`, which probably isn't even necessary.  Do
the same thing with units depending on the local FreeIPA service.

[1] https://www.digitalocean.com/community/tutorials/how-to-create-flexible-services-for-a-coreos-cluster-with-fleet-unit-files