IPA containers:  master gets systemd service; client removed
Expecting that FreeIPA can't be self-hosted in k8s because of not only
chicken-and-egg bootstrapping but also chicken-and-egg booting
conundrums, start `ipa` container from a systemd service similar to
its old fleet unit-file.

Remove the ipaclient container and use the existing certmonger (and
running `ipa-getcert`) in the ipa container.

Adjust the `ipa_certmonger_certs.yaml` and `ipa_client_certs.yaml`
task includes to be more general.

The old freeipa Docker client is now only used for the toolbox image.