Rework bootstrapping, part 1:  bootstrap etcd/flanneld with temp certs
This starts a restructure of bootstrapping etcd/flanneld and IPA.

There's a chicken-and-egg problem in the bootstrap sequence:  To make
the IPA DNS server available across the cluster on internal, fixed
IPs, flannel must be running.  However, flannel depends on etcd, which
must be secured with TLS certs that are ultimately generated by IPA,
and which needs DNS SRV records to join members to the etcd cluster.

This restructure will work around the problem by creating a temporary
SSL CA and temporary DNS server in order to get etcd and flanneld
running.  This commit accomplishes those.

Then, a dedicated Docker network will be added with a separate flannel
instance for that network.  This is required to give the IPA DNS
servers a fixed IP address by making the host network CIDR /30.  Then
FreeIPA may be installed as before.

To clean up, SRV records will be added in FreeIPA, DNS will be pointed
to IPA servers, and new certs will be generated with IPA and etcd
restarted.