Back etcd3 version down to 3.1.5
In etcd bug 8445, they explain that certs need a domain name as well
as a hostname in the DNS SAN when using SRV records for cluster
discovery.

This would be more elegantly fixed by doing that, of course.  Should
be easy with cfssl, but unknown if FreeIPA will allow it.

https://github.com/coreos/etcd/pull/8445